+++ control-plane/subcommand/server-acl-init/connect_inject.go
@@ -1,7 +1,6 @@
 package serveraclinit
 
 import (
-	"errors"
 	"fmt"
 
 	"github.com/hashicorp/consul-k8s/control-plane/namespaces"
@@ -18,13 +17,11 @@ const defaultKubernetesHost = "https://kubernetes.default.svc"
 
 // configureConnectInject sets up auth methods so that connect injection will
 // work.
-func (c *Command) configureConnectInjectAuthMethod(consulClient *api.Client) error {
-
-	authMethodName := c.withPrefix("k8s-auth-method")
+func (c *Command) configureConnectInjectAuthMethod(consulClient *api.Client, authMethodName string) error {
 
 	// Create the auth method template. This requires calls to the
 	// kubernetes environment.
-	authMethodTmpl, err := c.createAuthMethodTmpl(authMethodName)
+	authMethodTmpl, err := c.createAuthMethodTmpl(authMethodName, true)
 	if err != nil {
 		return err
 	}
@@ -68,6 +65,7 @@ func (c *Command) configureConnectInjectAuthMethod(consulClient *api.Client) err
 		return err
 	}
 
+	c.log.Info("creating inject binding rule")
 	// Create the binding rule.
 	abr := api.ACLBindingRule{
 		Description: "Kubernetes binding rule",
@@ -76,74 +74,20 @@ func (c *Command) configureConnectInjectAuthMethod(consulClient *api.Client) err
 		BindName:    "${serviceaccount.name}",
 		Selector:    c.flagBindingRuleSelector,
 	}
-
-	// Binding rule list api call query options
-	queryOptions := api.QueryOptions{}
-
-	// Add a namespace if appropriate
-	// If namespaces and mirroring are enabled, this is not necessary because
-	// the binding rule will fall back to being created in the Consul `default`
-	// namespace automatically, as is necessary for mirroring.
-	if c.flagEnableNamespaces && !c.flagEnableInjectK8SNSMirroring {
-		abr.Namespace = c.flagConsulInjectDestinationNamespace
-		queryOptions.Namespace = c.flagConsulInjectDestinationNamespace
-	}
-
-	var existingRules []*api.ACLBindingRule
-	err = c.untilSucceeds(fmt.Sprintf("listing binding rules for auth method %s", authMethodName),
-		func() error {
-			var err error
-			existingRules, _, err = consulClient.ACL().BindingRuleList(authMethodName, &queryOptions)
-			return err
-		})
-	if err != nil {
-		return err
-	}
-
-	// If the binding rule already exists, update it
-	// This updates the binding rule any time the acl bootstrapping
-	// command is rerun, which is a bit of extra overhead, but is
-	// necessary to pick up any potential config changes.
-	if len(existingRules) > 0 {
-		// Find the policy that matches our name and description
-		// and that's the ID we need
-		for _, existingRule := range existingRules {
-			if existingRule.BindName == abr.BindName && existingRule.Description == abr.Description {
-				abr.ID = existingRule.ID
-			}
-		}
-
-		// This will only happen if there are existing policies
-		// for this auth method, but none that match the binding
-		// rule set up here in the bootstrap method.
-		if abr.ID == "" {
-			return errors.New("unable to find a matching ACL binding rule to update")
-		}
-
-		err = c.untilSucceeds(fmt.Sprintf("updating acl binding rule for %s", authMethodName),
-			func() error {
-				_, _, err := consulClient.ACL().BindingRuleUpdate(&abr, nil)
-				return err
-			})
-	} else {
-		// Otherwise create the binding rule
-		err = c.untilSucceeds(fmt.Sprintf("creating acl binding rule for %s", authMethodName),
-			func() error {
-				_, _, err := consulClient.ACL().BindingRuleCreate(&abr, nil)
-				return err
-			})
-	}
-	return err
+	return c.createConnectBindingRule(consulClient, authMethodName, &abr)
 }
 
-func (c *Command) createAuthMethodTmpl(authMethodName string) (api.ACLAuthMethod, error) {
+// createAuthMethodTmpl sets up the auth method template based on the connect-injector's service account
+// jwt token. It is common for both the connect inject auth method and the component auth method
+// with the option to add namespace specific configuration to the auth method template via `useNS`.
+func (c *Command) createAuthMethodTmpl(authMethodName string, useNS bool) (api.ACLAuthMethod, error) {
 	// Get the Secret name for the auth method ServiceAccount.
 	var authMethodServiceAccount *apiv1.ServiceAccount
-	saName := c.withPrefix("connect-injector")
-	err := c.untilSucceeds(fmt.Sprintf("getting %s ServiceAccount", saName),
+	serviceAccountName := c.withPrefix("auth-method")
+	err := c.untilSucceeds(fmt.Sprintf("getting %s ServiceAccount", serviceAccountName),
 		func() error {
 			var err error
-			authMethodServiceAccount, err = c.clientset.CoreV1().ServiceAccounts(c.flagK8sNamespace).Get(c.ctx, saName, metav1.GetOptions{})
+			authMethodServiceAccount, err = c.clientset.CoreV1().ServiceAccounts(c.flagK8sNamespace).Get(c.ctx, serviceAccountName, metav1.GetOptions{})
 			return err
 		})
 	if err != nil {
@@ -175,14 +119,14 @@ func (c *Command) createAuthMethodTmpl(authMethodName string) (api.ACLAuthMethod
 	// a secret of type ServiceAccountToken.
 	if saSecret == nil {
 		return api.ACLAuthMethod{},
-			fmt.Errorf("found no secret of type 'kubernetes.io/service-account-token' associated with the %s service account", saName)
+			fmt.Errorf("found no secret of type 'kubernetes.io/service-account-token' associated with the %s service account", serviceAccountName)
 	}
 
 	kubernetesHost := defaultKubernetesHost
 
 	// Check if custom auth method Host and CACert are provided
-	if c.flagInjectAuthMethodHost != "" {
-		kubernetesHost = c.flagInjectAuthMethodHost
+	if c.flagAuthMethodHost != "" {
+		kubernetesHost = c.flagAuthMethodHost
 	}
 
 	// Now we're ready to set up Consul's auth method.
@@ -197,8 +141,9 @@ func (c *Command) createAuthMethodTmpl(authMethodName string) (api.ACLAuthMethod
 		},
 	}
 
-	// Add options for mirroring namespaces
-	if c.flagEnableNamespaces && c.flagEnableInjectK8SNSMirroring {
+	// Add options for mirroring namespaces, this is only used by the connect inject auth method
+	// and so can be disabled for the component auth method.
+	if useNS && c.flagEnableNamespaces && c.flagEnableInjectK8SNSMirroring {
 		authMethodTmpl.Config["MapNamespaces"] = true
 		authMethodTmpl.Config["ConsulNamespacePrefix"] = c.flagInjectK8SNSMirroringPrefix
 	}